+++ fatiando/gravmag/__init__.py
@@ -37,6 +37,8 @@ The inversion modules use the forward modeling models and the
 The processing modules offer tools to prepare potential field data before or
 after modeling.
 
+* :mod:`~fatiando.gravmag.normal_gravity`: Compute normal gravity and
+  reductions.
 * :mod:`~fatiando.gravmag.eqlayer`: Equivalent layer processing
 * :mod:`~fatiando.gravmag.transform`: Space domain potential field
   transformations, like upward continuation