+++ doc/changelog.rst
@@ -10,6 +10,10 @@ Version (future)
 
 **Changes**:
 
+* **New module** :ref:`fatiando.gravmag.normal_gravity
+  <fatiando_gravmag_normal_gravity>` to calculate normal gravity (the gravity
+  of reference ellipsoids).
+  (`PR 133 <https://github.com/fatiando/fatiando/pull/133>`_)
 * Using `versioneer <https://github.com/warner/python-versioneer>`__ to manage
   version numbers. Access the version number + git commit hash from
   ``fatiando.__version__``.